enhance image segmentation

+++ get_segments_urbancross_enhanced_urbancross_reversed.py
@@ -103,14 +103,16 @@ def show_masks_mine(anns, ori_img, img_path, description):
         print(f"Kept {path} with similarity {similarity}")
 
     # 删除其他分割
-    for _, path in sorted(similarities, key=lambda x: x[0], reverse=True)[10:]:
-        os.remove(path)
+    for similarity, path in sorted(similarities, key=lambda x: x[0], reverse=True)[10:]:
+        if os.path.exists(path):
+            os.remove(path)
+
 
 if __name__ == "__main__":
-    # img_path = "/hpc2hdd/home/szhong691/zsr/projects/dataset/UrbanCross/image_target/Finland/images"
-    img_path = "/hpc2hdd/home/szhong691/zsr/projects/dataset/UrbanCross/image_target/Germany/images"
-    # df = pd.read_csv("/hpc2hdd/home/szhong691/zsr/projects/dataset/UrbanCross/image_target/Finland/instructblip_generation_finland_refine.csv")
-    df = pd.read_csv("/hpc2hdd/home/szhong691/zsr/projects/dataset/UrbanCross/image_target/Germany/instructblip_generation_germany_refine.csv")
+    img_path = "/hpc2hdd/home/szhong691/zsr/projects/dataset/UrbanCross/image_target/Finland/images"
+    # img_path = "/hpc2hdd/home/szhong691/zsr/projects/dataset/UrbanCross/image_target/Germany/images"
+    df = pd.read_csv("/hpc2hdd/home/szhong691/zsr/projects/dataset/UrbanCross/image_target/Finland/instructblip_generation_finland_refine.csv")
+    # df = pd.read_csv("/hpc2hdd/home/szhong691/zsr/projects/dataset/UrbanCross/image_target/Germany/instructblip_generation_germany_refine.csv")
 
     sam_checkpoint = "sam_vit_h_4b8939.pth"
     model_type = "vit_h"
@@ -119,7 +121,7 @@ if __name__ == "__main__":
     mask_generator = SamAutomaticMaskGenerator(sam)
 
     img_lists = df["image_name"]
-    for idx, row in tqdm(reversed(list(df.iterrows())), total=df.shape[0]):
+    for idx, row in tqdm(reversed(list(df.iterrows())[7000:]), total=df.shape[0]):
         image_name = row['image_name']
         description = row['description']  # 确保CSV中有描述的列
         image_path = os.path.join(img_path, image_name)
